+++ client/src/socket/signalingClient.ts
@@ -1,7 +1,5 @@
 import { io } from 'socket.io-client';
 
-type SocketId = string | undefined;
-
 const configuration = {
   iceServers: [
     {
@@ -21,70 +19,40 @@ const configuration = {
   ],
 };
 
-export default (
-  localStream: MediaStream | null,
-  remoteStreamMap: Map<SocketId, MediaStream>,
-  nickNameMap: Map<SocketId, string>,
-  stopWatchMap: Map<SocketId, RTCDataChannel>,
-) => {
-  const socket = io(import.meta.env.VITE_SIGNALING_SERVER_URL);
-  const peerConnectionMap = new Map<SocketId, RTCPeerConnection>();
+interface WebRTCData {
+  peerConnection: RTCPeerConnection;
+  remoteStream: MediaStream;
+  dataChannel: RTCDataChannel;
+  nickName: string;
+}
 
-  // ICE candidate 수신 및 추가
-  socket.on('setIceCandidate', (data) => {
+const signalingClient = (localStream: MediaStream, webRTCMap: Map<string, WebRTCData>) => {
+  const socket = io(import.meta.env.VITE_SIGNALING_SERVER_URL);
+
+  socket.on('setIceCandidate', (data: string) => {
     const { senderId, iceCandidate } = JSON.parse(data);
-    peerConnectionMap.get(senderId)!.addIceCandidate(new RTCIceCandidate(iceCandidate));
+    const { peerConnection } = webRTCMap.get(senderId)!;
+    if (peerConnection) {
+      peerConnection.addIceCandidate(new RTCIceCandidate(iceCandidate));
+    }
   });
 
-  socket.on('offerRequest', async (data) => {
+  socket.on('offerRequest', async (data: string) => {
     const { users } = JSON.parse(data);
     for (const oldId of users) {
-      // RTCPeerConnection 생성
       const peerConnection = new RTCPeerConnection(configuration);
-      peerConnectionMap.set(oldId, peerConnection);
 
       const dataChannel = peerConnection.createDataChannel('stopWatch');
-      let interval: NodeJS.Timeout;
-      dataChannel.onopen = () => {
-        console.log(`Data channel from ${oldId} is open`);
-        console.log(`Socket Id is ${socket.id}`);
 
-        // 채널이 열린 후에 메시지 전송 시작
-        interval = setInterval(() => {
-          console.log('interval');
-          dataChannel.send('22:56:00');
-        }, 1000);
-      };
-
-      dataChannel.onmessage = ({ data }) => {
-        console.log('Received:', data);
-      };
-
-      dataChannel.onclose = () => {
-        console.log('DataChannel closed');
-        clearInterval(interval);
-      };
-
-      stopWatchMap.set(oldId, dataChannel);
-
-      // 미디어 스트림 트랙 전송
-      if (!localStream) {
-        return;
-      }
       localStream.getTracks().forEach((track) => peerConnection.addTrack(track, localStream));
 
-      // offer 생성
       const offer = await peerConnection.createOffer();
 
-      socket.emit('sendOffer', { oldId, offer, newRandomId: localStorage.getItem('nickName') });
-
-      // LocalDescription 설정
       await peerConnection.setLocalDescription(offer);
 
-      // ICE candidate 전송
-      peerConnection.onicecandidate = (event) => {
-        if (event.candidate) {
-          socket.emit('sendIceCandidate', { targetId: oldId, iceCandidate: event.candidate });
+      peerConnection.onicecandidate = ({ candidate }) => {
+        if (candidate) {
+          socket.emit('sendIceCandidate', { targetId: oldId, iceCandidate: candidate });
         }
       };
 
@@ -94,69 +62,43 @@ export default (
           peerConnection.connectionState === 'failed' ||
           peerConnection.connectionState === 'closed'
         ) {
-          remoteStreamMap.delete(oldId);
-          peerConnectionMap.delete(oldId);
-          stopWatchMap.delete(oldId);
+          webRTCMap.delete(oldId);
           peerConnection.close();
         }
       };
+
+      socket.emit('sendOffer', { oldId, offer, newRandomId: localStorage.getItem('nickName') });
+
+      webRTCMap.set(oldId, { ...webRTCMap.get(oldId)!, peerConnection, dataChannel });
     }
   });
 
   socket.on('answerRequest', async (data) => {
     const { newId, offer, newRandomId } = JSON.parse(data);
 
-    nickNameMap.set(newId, newRandomId);
-
-    // RTCPeerConnection 생성
     const peerConnection = new RTCPeerConnection(configuration);
-    peerConnectionMap.set(newId, peerConnection);
 
     peerConnection.ondatachannel = ({ channel }) => {
-      // 채널 상태 핸들러 설정
-      channel.onopen = () => {
-        console.log(`Data channel from ${newId} is open`);
-        console.log(`Socket Id is ${socket.id}`);
-      };
-
-      channel.onmessage = (event) => {
-        console.log(`Message from ${newId}:`, event.data);
-      };
-
-      channel.onclose = () => {
-        console.log(`Data channel from ${newId} closed`);
-        stopWatchMap.delete(newId);
-      };
-
-      stopWatchMap.set(newId, channel);
+      // dataChannel 이벤트 발생 시점에 webRTCMap의 최신 데이터를 가져와서 dataChannel을 업데이트
+      webRTCMap.set(newId, { ...webRTCMap.get(newId)!, dataChannel: channel });
     };
 
-    // 미디어 스트림 트랙 수신
-    peerConnection.ontrack = (event) => {
-      remoteStreamMap.set(newId, event.streams[0]);
+    peerConnection.ontrack = ({ streams }) => {
+      // track 이벤트 발생 시점에 webRTCMap의 최신 데이터를 가져와서 remoteStream을 업데이트
+      webRTCMap.set(newId, { ...webRTCMap.get(newId)!, remoteStream: streams[0] });
     };
 
-    // RemoteDescription 설정
     await peerConnection.setRemoteDescription(new RTCSessionDescription(offer));
 
-    // 미디어 스트림 트랙 전송
-    if (!localStream) {
-      return;
-    }
     localStream.getTracks().forEach((track) => peerConnection.addTrack(track, localStream));
 
-    // Answer 생성
     const answer = await peerConnection.createAnswer();
 
-    socket.emit('sendAnswer', { newId, answer, oldRandomId: localStorage.getItem('nickName') });
-
-    // LocalDescription 설정
     await peerConnection.setLocalDescription(answer);
 
-    // ICE candidate 전송
-    peerConnection.onicecandidate = (event) => {
-      if (event.candidate) {
-        socket.emit('sendIceCandidate', { targetId: newId, iceCandidate: event.candidate });
+    peerConnection.onicecandidate = ({ candidate }) => {
+      if (candidate) {
+        socket.emit('sendIceCandidate', { targetId: newId, iceCandidate: candidate });
       }
     };
 
@@ -166,30 +108,32 @@ export default (
         peerConnection.connectionState === 'failed' ||
         peerConnection.connectionState === 'closed'
       ) {
-        remoteStreamMap.delete(newId);
-        peerConnectionMap.delete(newId);
-        stopWatchMap.delete(newId);
+        webRTCMap.delete(newId);
         peerConnection.close();
       }
     };
+
+    socket.emit('sendAnswer', { newId, answer, oldRandomId: localStorage.getItem('nickName') });
+
+    webRTCMap.set(newId, { ...webRTCMap.get(newId)!, peerConnection, nickName: newRandomId });
   });
 
   socket.on('completeConnection', async (data) => {
     const { oldId, answer, oldRandomId } = JSON.parse(data);
 
-    nickNameMap.set(oldId, oldRandomId);
+    const { peerConnection } = webRTCMap.get(oldId)!;
 
-    // RTCPeerConnection 완료
-    const peerConnection = peerConnectionMap.get(oldId)!;
-
-    // 미디어 스트림 트랙 수신
-    peerConnection.ontrack = (event) => {
-      remoteStreamMap.set(oldId, event.streams[0]);
+    peerConnection.ontrack = ({ streams }) => {
+      // track 이벤트 발생 시점에 webRTCMap의 최신 데이터를 가져와서 remoteStream을 업데이트
+      webRTCMap.set(oldId, { ...webRTCMap.get(oldId)!, remoteStream: streams[0] });
     };
 
-    // RemoteDescription 설정
     await peerConnection.setRemoteDescription(new RTCSessionDescription(answer));
+
+    webRTCMap.set(oldId, { ...webRTCMap.get(oldId)!, nickName: oldRandomId });
   });
 
-  return { peerConnectionMap, socket };
+  return socket;
 };
+
+export default signalingClient;
